LayoutTests/fast/dom/Geolocation/dataURL-watchPosition-expected.txt

PASS error.code is error.POSITION_UNAVAILABLE.
PASS error.message is "Origin does not have permission to use Geolocation service".
